ASEI Update
The todo tags have been removed since all have been implemented and
ASEI is in a working state according to design spec. The global declaration
of the ASEI object was also removed so that executional code is not happening
at the time of the include, but rather upon the method call from another
server-side or client-side request.

Some clean up was done as well. The depreciated functions were removed and the
code comments were updated on several methods. This by no mean completes the
code portion of the documentation, but is at least slightly better than prior.

The current state of ASEI is that is can take PUT, GET, POST, DELETE commands
and pass them to an extended apiObject. It can take input in JSON and POST
formats currently, but it not capturing the PUT or DELETE commands. It can 
respond in JSON, HTML, XML, or STREAM. The xml is still only working partialy.