Server Side Libraries
Server-side libraries were cleaned up and organized to better serve
the overall file structure. Part of this change was due to changes in
the boot process which required the workflow to fork. So it was decided
that further forms could be possible, and if that were the case the current
structe may not be scalable.

Boot Fork: A fork in the boot process was added to solve some issues
where we need some things to happen when starting a rest response, but
not to happen when loading a page, but continue to leave in-tact, the 
restful request structure and methods.

Basic HTML: Some basic HTML was added to serve as landing pages for
RESI during testing. Currently the process show a traditional GET
request as well as an ASEI page request using onServer.page();